Restructured folders (style classes), added non-form elements of register page
