+++ src/components/navbar/navbar.css
@@ -19,8 +19,3 @@ li{
     margin: 0 1rem;
     list-style-type: none;
 }
-
-a{
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-}
