--- src/pages/register.js
@@ -1,5 +1,0 @@
-const Register = () => {
-    return <h1>Register</h1>
-};
-
-export default Register;
