+++ src/style/bodyLink.js
@@ -1,0 +1,6 @@
+const bodyLink = {
+    color: "#547AA5",
+    textDecoration: "None"
+};
+
+export default bodyLink;
