+++ src/style/buttonPrimary.js
@@ -1,7 +1,9 @@
 const primaryButton = {
     color: "white",
     backgroundColor: "#547AA5",
-    textTransform: "uppercase"
+    textTransform: "uppercase",
+    textDecoration: "None",
+    padding: "0.5rem 1rem"
 }
 
 export default primaryButton;
