+++ src/style/buttonSecondary.js
@@ -1,7 +1,9 @@
 const secondaryButton = {
     color: "white",
     backgroundColor: "#4F5165",
-    textTransform: "uppercase"
+    textTransform: "uppercase",
+    textDecoration: "None",
+    padding: "0.5rem 1rem"
 }
 
 export default secondaryButton;
